Change calculation

+++ Software/Loom 4/FloDar_Field/FloDar_Field.ino
@@ -40,7 +40,7 @@ Loom_ADS1115 ads(manager);
 
 // Analog to Flowrate Conversion Formula
 float calcFlowRate(){
-    return ((ads.getAnalog(3) / MAXBITS) * MAXFLOWRATE);
+    return (((float)ads.getAnalog(3) / MAXBITS) * MAXFLOWRATE);
 }
 
 // Called when the interrupt is triggered
